# openapi_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PyDance Quantum-Secure API",
    "version": "2.0.0",
    "description": "A quantum-resistant web framework API with advanced security features"
  },
  "paths": {
    "/api/v1/users": {
      "get": {
        "summary": "List Users",
        "description": "Retrieve a paginated list of users",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {}
      },
      "post": {
        "summary": "Create User",
        "description": "Create a new user account",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "summary": "Get User",
        "description": "Retrieve a specific user by ID",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "User Login",
        "description": "Authenticate user and return access token",
        "tags": [
          "authentication"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/quantum/keypair": {
      "post": {
        "summary": "Generate Quantum Keypair",
        "description": "Generate a quantum-resistant cryptographic keypair",
        "tags": [
          "quantum"
        ],
        "parameters": [
          {
            "name": "algorithm",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "kyber"
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/quantum/authenticate": {
      "post": {
        "summary": "Quantum Authentication",
        "description": "Perform quantum-resistant authentication",
        "tags": [
          "quantum",
          "authentication"
        ],
        "parameters": [
          {
            "name": "identity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/system/health": {
      "get": {
        "summary": "System Health Check",
        "description": "Check system health and status",
        "tags": [
          "system"
        ],
        "parameters": [],
        "responses": {}
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "username",
          "email"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "ValidationError"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
      "QuantumKeypair": {
        "type": "object",
        "properties": {
          "algorithm": {
            "type": "string",
            "example": "kyber"
          },
          "public_key": {
            "type": "string",
            "description": "Base64-encoded public key"
          },
          "private_key": {
            "type": "string",
            "description": "Base64-encoded private key"
          }
        },
        "required": [
          "algorithm",
          "public_key",
          "private_key"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "quantumAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Quantum-Auth",
        "description": "Quantum-resistant authentication token"
      }
    }
  },
  "tags": [
    {
      "name": "authentication",
      "description": "Authentication and authorization endpoints"
    },
    {
      "name": "users",
      "description": "User management endpoints"
    },
    {
      "name": "quantum",
      "description": "Quantum-resistant security endpoints"
    },
    {
      "name": "system",
      "description": "System and monitoring endpoints"
    }
  ]
}